separate settings components into 2
