+++ frontend/components/settings/EditProfile.vue
@@ -1,0 +1,11 @@
+<template>
+  <div>test</div>
+</template>
+<script>
+export default {
+  data () {
+    return {
+    }
+  }
+}
+</script>
